add MacOS compat in Makefile, move core files to dedicated folder, rename files according to PEP8
