+++ locatorfilters/google_api_key_dialog.py
@@ -26,13 +26,13 @@ from PyQt5 import uic
 from PyQt5 import QtWidgets
 
 FORM_CLASS, _ = uic.loadUiType(os.path.join(
-    os.path.dirname(__file__), 'base_api_key_dialog.ui'))
+    os.path.dirname(__file__), 'google_api_key_dialog.ui'))
 
 
-class BaseApiKeyDialog(QtWidgets.QDialog, FORM_CLASS):
+class GoogleApiKeyDialog(QtWidgets.QDialog, FORM_CLASS):
     def __init__(self, parent=None):
         """Constructor."""
-        super(BaseApiKeyDialog, self).__init__(parent)
+        super(GoogleApiKeyDialog, self).__init__(parent)
         # Set up the user interface from Designer.
         # After setupUI you can access any designer object by doing
         # self.<objectname>, and you can use autoconnect slots - see
