+++ locatorfilters/nominatim.py
@@ -1,7 +1,10 @@
-from qgis.core import QgsProject, QgsLocatorResult, QgsRectangle, QgsCoordinateReferenceSystem, QgsCoordinateTransform
-from .basefilter import GeocoderFilter
-from ..networkaccessmanager import RequestsException
 import json
+
+from qgis.core import QgsProject, QgsLocatorResult, QgsRectangle, QgsCoordinateReferenceSystem, QgsCoordinateTransform
+
+from ..core.network_access_manager import RequestsException
+from ..core.base_filter import GeocoderFilter
+
 
 class NominatimFilter(GeocoderFilter):
